+++ packages/dart/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `getFileSize` method
 - Add `normalizeFile` method
+- Add `headFile` method
 
 ## 1.1.8 (2025-11-13)